+++ test/fixtures/with-copy/application.js
@@ -1,0 +1,3 @@
+import logoUrl from './logo.svg'
+
+console.log('hello', logoUrl)